xform_cpp/server/requirements.txt
Flask==0.9
numpy>=1.8.1
Pillow==2.7.0
